backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.1.2
blinker==1.9.0
click==8.2.1
ecdsa==0.19.1
fastapi==0.109.2
Flask==3.1.1
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.3.0
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.27.1
Werkzeug==3.1.3
